권한체크 및 저자정보 추가

+++ sparta_news/articles/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import get_object_or_404
 from .models import Article
 from django.core import serializers
 from rest_framework.response import Response
+from rest_framework.permissions import IsAuthenticated
 from rest_framework import status
 from rest_framework.views import APIView
 from .serializers import ArticleSerializer
@@ -13,9 +14,11 @@ class ArticleListAPIView(APIView):
         return Response(serializer.data)
 
     def post(self, request):
+        if not request.user.is_authenticated:
+            return Response({"error": "인증되지 않은 사용자입니다."}, status=status.HTTP_401_UNAUTHORIZED)
         serializer = ArticleSerializer(data=request.data)
         if serializer.is_valid(raise_exception=True):
-            serializer.save()
+            serializer.save(author=request.user)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
 
 class ArticleDetailAPIView(APIView):
@@ -28,6 +31,8 @@ class ArticleDetailAPIView(APIView):
         return Response(serializer.data)
 
     def put(self, request, pk):
+        if not request.user.is_authenticated:
+            return Response({"error": "인증되지 않은 사용자입니다."}, status=status.HTTP_401_UNAUTHORIZED)
         article = self.get_object(pk)
         serializer = ArticleSerializer(article, data=request.data, partial=True)
         if serializer.is_valid(raise_exception=True):
@@ -35,6 +40,8 @@ class ArticleDetailAPIView(APIView):
             return Response(serializer.data)
 
     def delete(self, request, pk):
+        if not request.user.is_authenticated:
+            return Response({"error": "인증되지 않은 사용자입니다."}, status=status.HTTP_401_UNAUTHORIZED)
         article = self.get_object(pk)
         article.delete()
         data = {"pk": f"{pk} is deleted."}
